prepare for another version of mbed_config
